+++ .gitignore
@@ -1,3 +1,3 @@
-lmd_matched
-lmd_matched_h5
-*.gz
+terminal.ipynb
+testing.ipynb
+data_demo.ipynb
