+++ README.md
@@ -1,1 +1,5 @@
 # cogs108-fa23
+
+This repo is for the data phase of our project.
+
+To begin, please navigate to src/load_data.ipynb. Read carefully and follow the instructions to replicate the data extraction. (We are using the Lakh Million Song Dataset)
