--- midi_methods.py
@@ -1,70 +1,0 @@
-import json
-from collections import defaultdict
-from sklearn import linear_model, model_selection
-import numpy as np
-import random
-import gzip
-import dateutil.parser
-import math
-import mido
-import pretty_midi
-import joblib
-import glob
-import matplotlib.pyplot as plt
-import seaborn as sns
-import matplotlib.gridspec
-import collections
-import os
-import tarfile
-
-
-def extract_midi_files(path='lmd_matched.tar.gz', output_folder='lmd_matched'):
-    """
-    Given a path to a tar.gz fiile, extract all compressed files
-
-    Parameters
-    ----------
-    path : str
-        Path to tar.gz file
-    output_folder : str
-        Destination folder to extract files to
-    
-    Returns
-    -------
-    None
-    """
-    f = tarfile.open(path)
-    f.extractall(output_folder)
-    f.close
-    return
-
-
-def compute_statistics(midi_file):
-    """
-    Given a path to a MIDI file, compute a dictionary of statistics about it
-    
-    Parameters
-    ----------
-    midi_file : str
-        Path to a MIDI file.
-    
-    Returns
-    -------
-    statistics : dict
-        Dictionary reporting the values for different events in the file.
-    """
-    # Some MIDI files will raise Exceptions on loading, if they are invalid.
-    # We just skip those.
-    try:
-        pm = pretty_midi.PrettyMIDI(midi_file)
-        # Extract informative events from the MIDI file
-        return {'n_instruments': len(pm.instruments),
-                'program_numbers': [i.program for i in pm.instruments if not i.is_drum],
-                'key_numbers': [k.key_number for k in pm.key_signature_changes],
-                'tempos': list(pm.get_tempo_changes()[1]),
-                'time_signature_changes': pm.time_signature_changes,
-                'end_time': pm.get_end_time(),
-                'lyrics': [l.text for l in pm.lyrics]}
-    # Silently ignore exceptions for a clean presentation (sorry Python!)
-    except Exception as e:
-        pass
