--- packages.txt
@@ -1,10 +1,0 @@
-scikit-learn
-numpy
-mido
-pretty_midi
-joblib
-matplotlib
-seaborn
-pandas
-tables
-
