+++ utils/utils.py
@@ -1,0 +1,15 @@
+import os
+import glob
+
+def get_all_files(basedir,ext='.h5') :
+    """
+    From a root directory, go through all subdirectories
+    and find all files with the given extension.
+    Return all absolute paths in a list.
+    """
+    allfiles = []
+    for root, dirs, files in os.walk(basedir):
+        files = glob.glob(os.path.join(root,'*'+ext))
+        for f in files :
+            allfiles.append( os.path.abspath(f) )
+    return allfiles
